+++ nodes/bucket-read.js
@@ -47,10 +47,16 @@ module.exports = function(RED) {
         queryParameters.set('items',limit > 1000 || limit < 0 ? 1000 : limit);
 
         // Query parameters to string
-        var queryParametersString = "";
+        let queryParametersString = "";
         queryParameters.forEach(function(value,key) {
             if (value) {
-                queryParametersString += key+"="+value+"&";
+                if ( Array.isArray(value) ) {
+                  for ( let i in value ) {
+                    queryParametersString += key+"="+value[i]+"&";
+                  }
+                } else {
+                  queryParametersString += key+"="+value+"&";
+                }
             }
         });
 
@@ -98,29 +104,37 @@ module.exports = function(RED) {
         RED.nodes.createNode(this, config);
 
         // get node
-        var node = this;
+        const node = this;
 
         // get server configuration
-        var server = RED.nodes.getNode(config.server);
+        const server = RED.nodes.getNode(config.server);
 
         // call bucket read on close
         node.on("input",function(msg, send, done) {
 
             let bucket = config.bucket || msg.bucket;
 
+            let tags = config.tags && Object.keys(config.tags).length !== 0 ? config.tags : msg.tags;
+
             const queryParameters = new Map();
             queryParameters.set('items',config.items || msg.items);
             queryParameters.set('agg',config.aggregation || msg.aggregation);
             queryParameters.set('agg_type',config.aggregationType || msg.aggregation_type);
             queryParameters.set('sort',config.sort || msg.sort);
-
+            for ( let key in tags ) {
+              let array = [];
+              for ( let i in tags[key] ) {
+                array.push(tags[key][i]);
+              }
+              queryParameters.set(key, array);
+            }
 
             // Timeframe filters
             let filter = config.filter || msg.filter;
-            var isFilterTime = true;
-            var isSimpleSorting = false;
-            var maxTs;
-            var minTs;
+            let isFilterTime = true;
+            let isSimpleSorting = false;
+            let maxTs;
+            let minTs;
 
             switch(filter) {
                 case "relative":
@@ -147,7 +161,7 @@ module.exports = function(RED) {
                     minTs = new Date(config.minTs || msg.min_ts).getTime();
                     break;
                 case "simple":
-                    //if selection of last N items, the query will be done as desc and sorted to asc after
+                    //if selection of last N items, the query will be done as desc and sorted to asc after, otherwise, result would not be last
                     if (queryParameters.get('sort') == "asc") {
                         queryParameters.set('sort','desc');
                         isSimpleSorting = true;
@@ -164,20 +178,20 @@ module.exports = function(RED) {
             }
 
             // limit < 0 will indicate all items matching the filter
-            var limit = queryParameters.get('items');
+            let limit = queryParameters.get('items');
             if (!limit) {
                 limit = -1;
             }
-            var result = [];
+            let result = [];
 
             readBucket(server, node, bucket, queryParameters, limit, result)
               .then(function(data) {
                 if (isSimpleSorting) { // sort last N items asc if needed
-                    result = data.sort(function(a,b) {
+                    data = data.sort(function(a,b) {
                         return a.ts - b.ts;
                     });
                 }
-                msg.payload = result;
+                msg.payload = data;
                 send(msg);
                 done();
               })