+++ nodes/bucket-write.js
@@ -7,16 +7,16 @@ module.exports = function(RED) {
         RED.nodes.createNode(this, config);
 
         // get node
-        var node = this;
+        const node = this;
 
         // get server configuration
-        var server = RED.nodes.getNode(config.server);
+        const server = RED.nodes.getNode(config.server);
 
         // call bucket write on message reception
         node.on("input", function(msg, _send, done) {
 
             let bucket = config.bucket || msg.bucket;
-            var value = config.value || msg.payload || msg.value;
+            let value = config.value || msg.payload || msg.value;
             if (typeof(value) === 'string') {
                 value = JSON.parse(value);
             }